Permitir decimales en rubrica y actualizacion de schema y seed

+++ backend/src/modules/reviews/reviews.validator.js
@@ -1,7 +1,8 @@
 // src/modules/reviews/reviews.validator.js
 import { z } from 'zod';
 
-const rubricField = z.coerce.number().int().min(1, 'Mínimo 1').max(5, 'Máximo 5');
+// CAMBIO: Se eliminó .int() para permitir decimales (ej. 3.5)
+const rubricField = z.coerce.number().min(1, 'Mínimo 1').max(5, 'Máximo 5');
 
 export const createAssignmentSchema = z.object({
   paperId:    z.coerce.number().int().positive(),
@@ -18,7 +19,7 @@ export const submitReviewSchema = z.object({
   assignmentId:        z.coerce.number().int().positive(),
   verdict:             z.enum(['ACCEPT', 'MINOR_CHANGES', 'MAJOR_CHANGES', 'REJECT']),
   originality:         rubricField,
-  methodologicalRigor: rubricField,
+  methodologicalRigor: rubricField, // Asegúrate que coincida con el nombre en tu DB/Frontend
   writingQuality:      rubricField,
   relevance:           rubricField,
   comments:            z.string().min(20, 'Los comentarios deben tener al menos 20 caracteres'),
